Moved translation to latin methods to parser
